Refactor: Preparing for compressed I/O
This refactoring changes quite a bit the TokenReader, TokenWriter traits:

1. A new method `TokenWriter::done` appears, which serves to
destructively access the contents of the TokenWriter. This proves more
useful when attempting to have a common API for uncompressed and
compressed I/O.

2. Instances of `TokenWriter` are not stacked anymore. Rather, there
is a single instance of `TokenWriter`, which when necessary return
`Guard` objects. These `Guard` objects serve to ensure that we have
properly read lists, tuples, etc. to their end. This design proves
more useful to be able to return proper exceptions in case of improper
input.

Smaller stuff around.