Resolves #4 Split Object{Getter, Setter, Method, Property} into two interfaces each.
Each of these interfaces had a field `key` and a field `computed`. If
`computed` was `true` (which is *extremely* uncommon and corresponds
to something along the lines of `{[1 + 1]: 100}` ), `key` was an
`Expression`. Otherwise, it was an `Identifier` or `Literal`.

Separating Object{Getter, ...} (the common case) from
BracketObject{Getter, ...} (the unusual case) makes the common case
more readable and more fuzzing-friendly, at no cost to file size.