Porting to Shift-style AST
The specifications have changed from a ESTree/Babel-style AST to a stricter Shift-style AST designed to simplify runtime verification
that an AST is well-formed.

See https://github.com/binast/ecmascript-binary-ast/blob/spec-draft/spec-draft.md for the current draft of the AST. This patch
implements the new AST.

It does *not* implement scope checking yet, as these are not specified at the time of this writing.