Issue #71 Generate the encoder and decoder at build-time
With this patch, we don't need to go through JSON to encode from/decode to a strongly typed AST. We still need JSON for parsing/pretty-printing, though.

While I haven't checked, in addition to avoiding all the weakly-typed parts of the code, this should speed things up quite a bit.