Resolves #176 - Getting rid of TokenWriter::{list, tagged_tuple} and the companion code in ast.rs
This should make future work on entropy a bit simpler, improve performance and this has already helped locate a few bugs in the generic implementation of encoding.

This will need to be followed up by a corresponding work on the TokenReader, once we have a clear idea of exactly what needs to be done.