Implemented binjs_generate_prediction_tables
This change adds a new command-line tool, binjs_generate_prediction_tables,
which analyzes a bunch of JS source files and outputs binary files containing
the prediction tables for the group. Once we have implemented v2 of the `entropy`
encoder, these prediction tables may then be used for encoding/decoding with
that format.