Utility: ContentInfo::get_mut_b
When parsing the [[content]] section of an Entropy 0.2+ file, we do not know for sure whether
all streams are present. This utility method let us access the individual fields of ContentInfo
(which represent streams in this case) by name, so as to update the ContentInfo data structure
as we encounter the stream definitions.