Entropy 0.2: Adapting dictionary
We add one type of dictionaries: the IndexedTable. We use IndexedTables to simply store values and recover them by index. In Entropy 0.2, we use this to create content streams for user-extensible values. These content streams are
streams of indices (i.e. varnums) that refer to either an IndexedTable (the static dictionary) or a local vector of values (the prelude dictionary).