Issue #209 - Implementing window predictions
This patch adds support for a mechanism of entropy prediction not by path in the AST, but by a sliding window. With this mechanism, we attempt to predict the most likely values for a backreference. May be used to encode
user-extensible values, along with a prelude dictionary. As of this writing, this is only used for building dictionaries and getting statistics, not to actually encode anything in the compressed file.