Issue #296 - Extending the webidl with a secondary file.
This patch adds support for extending es6.webidl with an additional file es6-extension.webidl. The role of that file is to modify the AST in a manner that is useful for *compression* but has no impact on visible *syntax and
semantics* of the language. In other words, es6.webidl is meant to be part of the BinAST language specification, while es6-extension.webidl is meant to be part of the compression format.