Resolve #302 - Reserve a number of indices to reference values recently used in content streams.
Experiments indicate that, in the case of the identifier content stream, reserving numbers [1, 8]
to represent the latest (respectively second-latest, etc.) value encountered increases compression
of that stream by 6+%. Other streams are affected negatively by the same change.

This patch introduces a mechanism of backreferences for numbers [1,8] for identifier content streams,
in a manner that may, at a later stage, be customized further (possibly as a command-line option).