Resolves #304 - Dynamically collect references to shared dictionary into a dense interval.
This patch changes how `TableRef` instances are encoded as u32. Prior to this patch, shared
references are represented as their index in the shared dictionary. Since the shared dictionary
is large, but sparsely used, this causes varnums to be needlessly large and all over the place,
hence hard to compress: typically, most references to the shared dictionary end up costing two
unpredictable bytes, and neither byte is predictable by the stream compressor (currently, brotli).

With this patch, we change the mechanism, to make sure that most uses of references into
the shared dictionary are consecutive and short, which makes them easier to compress with
brotli & co.

The general idea is the following:

- we maintain a vector of all the items of the shared dictionary that have been used at least
   once in the file - call this vector tor `import`;
- `import` is initially empty;
- whenever we first use a value of the shared dictionary with index `N` in the shared dictionary:
  - use `N + import.len()` to reference it;
  - push this value at the end of the shared dictionary, at position `P`.
- whenever we *reuse* a value already added to `import`
  - use `P` to reference it.

We then shift the result by adding `window.len() + prelude.len() + 1`, as numbers closer to 0 are
  reserved for other values:

- special number `0` (unaffected by this patch);
- special numbers `[1, window.len()]` (unaffected by this patch);
- numbers `]window.len(), window.len() + prelude.len()]`, to reference values in the prelude dictionary.

The consequence of this revised numbering is that, after one access with a random unpredictable number `N`, any further access to a reference in the shared dictionary is performed through a shorter number `P`, which may either
be single byte, or two bytes with the first byte  in a small range of possible values (typically {0b10000000, 0b01000000, 0b11000000}). Experimental results indicate that this improves compression for all streams.