ContentInfo becomes UserExtensible
While `ContentInfo` was initially design to gather statistics on contents, its main use is now to contain user-extensible data.

This patch:
- removes the fields that are not associated to user-extensible data;
- renames ContentInfo to UserExtensible.

Also, some of the statistics-gathering that took place during compression, and that used the now removed fields, has been deactivated. Since we're not using it, it's not really useful to keep at hand.