Remove lifetime from Serialization trait
In all cases `Serialization` already needs only a borrowed value, similarly to `serde::Serialize`, or an owned value of a primitive, which is easy and cheap to get from a borrow.

Hence, the definition can get simpler by always assuming a borrowed value of any type instead of requiring implementations with an explicit lifetime param.