AST Support for nodes that change the dictionary
This patch introduces a new annotation Â`ScopedDictionar=field` in the webidl. If an interface is decorated with this annotation, the typed AST will return the dictionary name held in `field` whenever we call
`Node::scoped_dictionary`. We will use this in a followup patch to perform actual changes of dictionaries for sub-asts, to handle e.g. JSON.