Issue 338 - Compare compression against a large sample
This patch introduces examples/sample_directory.rs. This binary takes as argument a directory name
and a sampling factor `f` for the dictionary. From the directory, `f`% of the files are then used
to build a dictionary, while the rest of the files are compressed.

Display the compression results and, optionally, the list of worst offenders.