Move BinaryAST -> Shift JSON conversion to JS side
This is the next step of #342.

Unlike the previous PR (#346), this not only tries to do 1:1 translation from Rust code to JavaScript, but also somewhat simplifies it where possible:

 - Ignores extra properties rather than trying to delete them, since Shift codegen will ignore them as well.
 - Merges all function branches into a catch-all regex.

Also one annoying complication is that JS conversion can't check the interface name it's converting into, so I had to hardcode some parents where `Block` and `VariableDeclaration` are not used as statements. These places are commented and hopefully won't be a big issue to maintain.