Fix performance of Deserialize
serde-derive implementation of tagged deserialisation is pretty inefficient (>2x slower than regular) and buffers entire input into an intermediate internal AST before searching for the tag key (see https://github.com/serde-rs/serde/issues/1495 for details).

It needs this to support generic cases where the tag key happens to be in arbitrary position of the input, but input is a non-seekable stream.

We use this deserialisation only in a well-defined context where we have control over both parts of the communication, so we can require `type` key to be the first in the JSON and avoid any buffering whatsoever to get the best possible performance.